10/29/2024 2nd updates
added the updates from the work with Kate from 10/25/2024 into the main branch of these code.  In addition to the other commit that were added today which updated the issue when there is not a time in the data this update deals with focuses on the transect min/ max times, and the short and long transect analyte min/ max.