Updated rawdataprep function
The raw data prep function needed to be updated to reflect the changes made to the MEM software as it stands now there the data are able reading in smoothly from the example code sent to me from Kate Haile on 2/3/2025.  There we also some changes made to the ts_plot function but that function is not at all finalized at this moment.  

Hopefully we are able to send along data to SAIC using this example for further development of the GMAP data dashboard with GMAP 2.0