ts_plot and MA_mapping function update
I got the ts_plot function working with both the wind directions figure and the time series.  Also got the MA_mapping function pretty wrapped up. there is still not documentation for any of the pieces at this point but that will likely occur once the core of the functions are completed.