Major updates
Major updates were made to the GMAPR package.  Functions to address pesudo-replication were developed for the Syft and picarro instruments. Also substantial edits were made to the MA and ST mapping functions to add additional context: including clarification of legends, kinds of landscape imagery available for use (Landsat and NAIP), and others.  Overall after some final tweaks we will likely be ready for a soft run on this code and data pipeline process. eventually code will need to be linted and annotated for full comprehensions but we are still a few steps away from that process.