lots of updates
Had a bunch of minor updates here and there, with the most major change being the pieces of the pipeline around and I believe the pesudoreplication error is likely handled at this time.  there will be another update likely at the end of this week (May 2, 2025) that will hopefully be the final update until a solid dry run is implemented.  We shall see..........